Split optimization passes into separate files
