+++ Makefile
@@ -14,7 +14,7 @@ CORE_SRC = src/main.c src/cli.c src/lexer.c src/ast_expr.c src/ast_stmt.c src/as
            src/preproc_macros.c src/preproc_expr.c src/preproc_file.c
 
 # Optional optimization sources
-OPT_SRC = src/opt.c
+OPT_SRC = src/opt.c src/opt_constprop.c src/opt_fold.c src/opt_dce.c
 # Additional sources can be specified by the user
 EXTRA_SRC ?=
 # Final source list
